requirements.txt
vcrpy==5.1.0
requests==2.31.0
